+++ packages/contracts-periphery/src/config/deploy.ts
@@ -5,14 +5,13 @@ import { ethers } from 'ethers'
  */
 export interface DeployConfig {
   /**
-   * Initial RetroReceiver owner.
+   * Dedicated Deterministic Deployer address (DDD).
+   * When deploying authenticated deterministic smart contracts to the same address on various
+   * chains, it's necessary to have a single root address that will initially own the contract and
+   * later transfer ownership to the final contract owner. We call this address the DDD. We expect
+   * the DDD to transfer ownership to the final contract owner very quickly after deployment.
    */
-  retroReceiverOwner: string
-
-  /**
-   * Initial Drippie owner.
-   */
-  drippieOwner: string
+  ddd: string
 }
 
 /**
@@ -24,10 +23,7 @@ const configSpec: {
     default?: any
   }
 } = {
-  retroReceiverOwner: {
-    type: 'address',
-  },
-  drippieOwner: {
+  ddd: {
     type: 'address',
   },
 }